Avoid slurping directories (fixes brightin/pottery#9)
This adds an additional check to the `pottery.scan/source-file?`
predicate fn to verify the file given is actually a file, and not a
directory.

It also adds a directory named `folder.clj` to the `test-resources`
folder, so that the `potter.scan-test/scan-files-test` test fails if
the fix is not applied.